Fixes for an exception and add function paramter types
In creating vscode.Range from moving beyond the first/last line or
first/last column, the corresponding variable does not have a valid
value (is undefined). This causes an exception when creating
vscode.Range. Not creating a range object is fine in that case.

Also add parameters types for overriding functions.