api: make switchToDocument a no-op for same doc
Switching to a document always leads to VS Code focusing on the document
in a **non-diff** editor, so if we focus on the current document while
editing it in a diff-editor, it will exit the diff, which is undesired.
This change makes switchToDocument a no-op when switching to the current
document, avoiding this problem.

Fixes #268.