commands: add .seek.leap (wip)
Inspired by https://github.com/ggandor/leap.nvim, `.seek.leap`
leaps to the first position on the screen that matches the two given
characters, with the ability to further jump to other labels.

`s` and `a-s` are rebound to call `.seek.leap` when all current
selections are empty.

Unfortunately, due to limitations in VS Code (/ lack of proper hacks
in Dance), the character sent directly after the leap will be
ignored.