Allows `seek` to accept more than one keypress.
The new argument `inputLength` (which defaults to 1) determines how many
characters the `seek` command accepts. This uses the newly implemented
`keyCount` argument (which also defaults to 1) to accept multiples
keypresses.

The most immediate use for this is to implement the equivalent of
vim-sneak. Sneak accepts two key presses, and like `t`, sneak jumps the
cursor right before the next match of the input.

This also includes some shortcut commands for easily sneaking with the
`s` key in normal mode using select.orSneak.