fix: jump vertically by page folding bug
Before this change, the jump vertically by page / halfPage fails when 
the first visible range is near a folded region. This solution 
considers all visible ranges, rather than just the first to calculate 
the page height. Alternatively, there can be a minimum number of lines 
to consider a "page", or the largest unfolded region could be used, or 
the difference between the first region's start and the last region's 
end.