feat: Add the rest of match mode to Helix keymap (#386)
This pull request adds `ms`, `mr`, and `md` to the Helix keymap, along
with the following fixes:
* Line numbers are now absolute in `helix/insert` and `helix/select`,
relative in `helix/normal`.
* Unimpaired mappings `[` and `]` are added back to the Helix keymap,
they were missing for some reason.

All these keybinds use existing Dance commands, the only limitation
being that `md` will replace the nearest pair instead of asking for the
pair to delete. It may be more efficient to implement these as commands,
though I understand that the core is still supposed to be Kakoune-based.