commands: add Helix goto word commands (#370)
Adds two new commands: 'Goto command' and 'Goto command (extend)'.

The commands add 2 character labels to every visible word of length >=
2. Then, they wait for user input. After receiving two characters of
input, the regular command jumps to the selected word, and the extend
command extends from the primary selection to the current word.

I don't know how to add keybinds for this so I haven't done so - both
commands should be activated by `gw` in helix.

This command is also affected by the bug described in #369.

Resolves: #339

Co-authored-by: jgdhs27 <jgdhs27@gmail.com>